프로필 전환 UI를 ProfileSwitcher 컴포넌트로 추출하고 사이드바 아바타를 Avatar로 교체한다
