+++ apps/web/src/lib/components/SidebarNavigation.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
   import { page } from '$app/state';
-  import { profileHandleSchema } from '@kosmo/core/validation';
-  import { createFragment, createMutation } from '@mearie/svelte';
+  import { createFragment } from '@mearie/svelte';
   import { graphql } from '$mearie';
+  import Avatar from '$lib/components/Avatar.svelte';
+  import ProfileSwitcher from '$lib/components/ProfileSwitcher.svelte';
   import { formatCount, getProfileInitial } from '$lib/utils/profile';
-  import type { DataOf } from '@mearie/svelte';
   import type { SidebarNavigation_query$key } from '$mearie';
 
   type Props = {
@@ -19,6 +19,7 @@
 
   const sidebarNavigationFragment = graphql(`
     fragment SidebarNavigation_query on Query {
+      ...ProfileSwitcher_query
       currentSession {
         selectedProfile {
           id
@@ -32,58 +33,10 @@
         id
         profiles {
           id
-          handle
-          displayName
-          followersCount
-          followingCount
         }
       }
     }
   `);
-
-  const selectSidebarProfileMutation = graphql(`
-    mutation SelectSidebarProfileMutation($id: ID!) {
-      selectProfile(input: { id: $id }) {
-        __typename
-        ... on SelectProfileSuccess {
-          profile {
-            id
-            handle
-            displayName
-            followersCount
-            followingCount
-          }
-        }
-        ... on NotFoundError {
-          message
-        }
-      }
-    }
-  `);
-
-  const createSidebarProfileMutation = graphql(`
-    mutation CreateSidebarProfileMutation($handle: String!) {
-      createProfile(input: { handle: $handle }) {
-        __typename
-        ... on CreateProfileSuccess {
-          profile {
-            id
-            handle
-            displayName
-            followersCount
-            followingCount
-          }
-        }
-        ... on ConflictError {
-          message
-          field
-        }
-      }
-    }
-  `);
-
-  type SidebarNavigationData = DataOf<typeof sidebarNavigationFragment>;
-  type ProfileSummary = NonNullable<SidebarNavigationData['me']>['profiles'][number];
 
   let {
     query = null,
@@ -126,15 +79,6 @@
   ];
 
   const sidebarNavigation = createFragment(sidebarNavigationFragment, () => query);
-  const [selectSidebarProfile] = createMutation(selectSidebarProfileMutation);
-  const [createSidebarProfile] = createMutation(createSidebarProfileMutation);
-  let profileError = $state<string | null>(null);
-  let profileCreationError = $state<string | null>(null);
-  let profileCreationOpen = $state(false);
-  let newProfileHandle = $state('');
-  let profileActionLoading = $state(false);
-  let selectedProfileOverride = $state<ProfileSummary | null>(null);
-  let profilesOverride = $state<ProfileSummary[] | null>(null);
 
   const isActive = (item: (typeof navItems)[number]) => {
     if (page.url.pathname !== item.href) {
@@ -148,226 +92,23 @@
     return true;
   };
 
-  const getProfileHandleError = (handle: string) => {
-    if (!handle) {
-      return '프로필 핸들을 입력해주세요.';
-    }
-
-    const result = profileHandleSchema.safeParse(handle);
-
-    return result.success
-      ? null
-      : (result.error.issues[0]?.message ?? '프로필 핸들 형식을 확인해주세요.');
-  };
-
-  const creatingOrSwitching = $derived(profileActionLoading || loading);
-  const sidebarProfiles = $derived(profilesOverride ?? sidebarNavigation.data?.me?.profiles ?? []);
+  const hasProfiles = $derived((sidebarNavigation.data?.me?.profiles?.length ?? 0) > 0);
   const sidebarActiveProfile = $derived(
-    selectedProfileOverride ?? sidebarNavigation.data?.currentSession?.selectedProfile ?? null,
+    sidebarNavigation.data?.currentSession?.selectedProfile ?? null,
   );
-
-  const openProfileSwitcher = () => {
-    switcherOpen = !switcherOpen;
-    if (!switcherOpen) {
-      profileCreationOpen = false;
-      profileCreationError = null;
-    }
-  };
-
-  const chooseProfile = async (id: string) => {
-    if (sidebarActiveProfile?.id === id || creatingOrSwitching) {
-      return;
-    }
-
-    profileError = null;
-    profileActionLoading = true;
-
-    try {
-      const data = await selectSidebarProfile({ id });
-
-      if (data.selectProfile.__typename !== 'SelectProfileSuccess') {
-        profileError = data.selectProfile.message;
-        return;
-      }
-
-      selectedProfileOverride =
-        sidebarProfiles.find((profile: ProfileSummary) => profile.id === id) ??
-        data.selectProfile.profile;
-      switcherOpen = false;
-      profileCreationOpen = false;
-      onProfileStateChanged();
-    } catch (error) {
-      profileError = error instanceof Error ? error.message : '프로필을 전환하지 못했습니다.';
-    } finally {
-      profileActionLoading = false;
-    }
-  };
-
-  const createAndSelectProfile = async (event: SubmitEvent) => {
-    event.preventDefault();
-
-    const handle = newProfileHandle.trim();
-
-    if (creatingOrSwitching) {
-      return;
-    }
-
-    const handleError = getProfileHandleError(handle);
-    if (handleError) {
-      profileCreationError = handleError;
-      return;
-    }
-
-    profileCreationError = null;
-    profileError = null;
-    profileActionLoading = true;
-
-    try {
-      const created = await createSidebarProfile({ handle });
-      if (created.createProfile.__typename !== 'CreateProfileSuccess') {
-        profileCreationError = created.createProfile.message;
-        return;
-      }
-
-      const createdProfile = created.createProfile.profile;
-
-      newProfileHandle = '';
-      profileCreationOpen = false;
-
-      const selected = await selectSidebarProfile({ id: created.createProfile.profile.id });
-      if (selected.selectProfile.__typename !== 'SelectProfileSuccess') {
-        profileError = selected.selectProfile.message;
-        return;
-      }
-
-      profilesOverride = [...sidebarProfiles, createdProfile];
-      selectedProfileOverride = createdProfile;
-      switcherOpen = false;
-      profileCreationOpen = false;
-      onProfileStateChanged();
-    } catch (error) {
-      profileCreationError =
-        error instanceof Error ? error.message : '프로필을 생성하지 못했습니다.';
-    } finally {
-      profileActionLoading = false;
-    }
-  };
 </script>
-
-<!-- 프로필 전환 드롭다운. 접힌 레일/풀 사이드바 양쪽에서 위치만 바꿔 재사용한다. -->
-{#snippet switcherMenu(idSuffix: string)}
-  {@const activeProfile = sidebarActiveProfile}
-  <div
-    class="flex w-[280px] flex-col gap-0.5 rounded-[14px] border border-[#eaeaea] bg-white p-1.5 shadow-[0_6px_20px_rgba(0,0,0,0.16)]"
-    role="menu"
-    aria-label="프로필 전환"
-  >
-    {#each sidebarProfiles as profile}
-      {@const selected = activeProfile?.id === profile.id}
-      <button
-        class={`flex w-full items-center gap-2.5 rounded-[10px] p-2 text-left transition disabled:opacity-50 ${selected ? 'bg-[#f6f6f6]' : 'hover:bg-[#f8f8f8]'}`}
-        type="button"
-        role="menuitemradio"
-        aria-checked={selected}
-        disabled={creatingOrSwitching}
-        onclick={() => chooseProfile(profile.id)}
-      >
-        <span
-          class={`flex shrink-0 items-center justify-center rounded-full bg-[#fce79a] font-bold text-[#111111] ${selected ? 'size-12 text-base' : 'size-8 text-sm'}`}
-          aria-hidden="true"
-        >
-          {getProfileInitial(profile.displayName, profile.handle)}
-        </span>
-        <span class="flex min-w-0 flex-1 flex-col gap-px overflow-hidden">
-          <span class="truncate text-sm font-bold text-[#111111]">{profile.displayName}</span>
-          <span class="truncate text-xs text-[#777777]">@{profile.handle}</span>
-        </span>
-        {#if selected}
-          <span class="text-[15px] font-bold text-[#111111]" aria-hidden="true">✓</span>
-        {/if}
-      </button>
-    {/each}
-
-    <div class="my-0.5 h-px w-full bg-[#eaeaea]"></div>
-
-    {#if profileCreationOpen}
-      <form
-        class="flex flex-col gap-1 p-1"
-        aria-label="새 프로필 만들기"
-        onsubmit={createAndSelectProfile}
-      >
-        <label class="sr-only" for={`profile-handle-${surface}-${idSuffix}`}>프로필 핸들</label>
-        <div class="flex gap-2">
-          <input
-            id={`profile-handle-${surface}-${idSuffix}`}
-            class="min-w-0 flex-1 rounded-lg border border-[#d4d4d8] px-3 py-2 text-sm outline-none transition placeholder:text-[#a1a1aa] focus:border-[#111111]"
-            name="handle"
-            autocomplete="off"
-            aria-describedby={`profile-handle-help-${surface}-${idSuffix}`}
-            aria-invalid={profileCreationError ? 'true' : undefined}
-            placeholder="새 프로필 핸들"
-            disabled={creatingOrSwitching}
-            bind:value={newProfileHandle}
-          />
-          <button
-            class="rounded-lg bg-[#fce79a] px-3 py-2 text-sm font-bold text-[#111111] transition hover:bg-[#f9dc6d] disabled:opacity-50"
-            type="submit"
-            disabled={creatingOrSwitching}
-          >
-            만들기
-          </button>
-        </div>
-        <p
-          id={`profile-handle-help-${surface}-${idSuffix}`}
-          class="px-1 text-xs leading-4 text-[#777777]"
-        >
-          영문, 숫자, 밑줄(_)만 사용할 수 있어요.
-        </p>
-        {#if profileCreationError}
-          <p class="px-1 text-xs leading-4 text-red-600">{profileCreationError}</p>
-        {/if}
-      </form>
-    {:else}
-      <button
-        class="flex w-full items-center gap-2.5 rounded-[10px] p-2 text-left transition hover:bg-[#f8f8f8] disabled:opacity-50"
-        type="button"
-        role="menuitem"
-        disabled={creatingOrSwitching}
-        onclick={() => {
-          profileCreationError = null;
-          profileCreationOpen = true;
-        }}
-      >
-        <span class="w-8 text-center text-lg font-bold text-[#111111]" aria-hidden="true">+</span>
-        <span class="text-sm font-medium text-[#111111]">새 프로필 추가</span>
-      </button>
-    {/if}
-  </div>
-{/snippet}
 
 <!-- md~lg 구간 접힌 아이콘 레일. 풀 사이드바와 스크립트/상태를 공유한다. -->
 {#snippet collapsedRail()}
   <div class="flex h-full w-full flex-col items-center gap-2 px-2 py-4">
-    <div class="relative">
-      <button
-        class="flex size-11 items-center justify-center rounded-full bg-text-primary text-lg font-bold text-white shadow-[1px_1px_2px_rgba(0,0,0,0.25)]"
-        type="button"
-        aria-label="프로필 목록"
-        aria-expanded={switcherOpen}
-        onclick={openProfileSwitcher}
-      >
-        {#if sidebarActiveProfile}
-          {getProfileInitial(sidebarActiveProfile.displayName, sidebarActiveProfile.handle)}
-        {:else}
-          ?
-        {/if}
-      </button>
-      {#if switcherOpen}
-        <div class="absolute left-[calc(100%+0.5rem)] top-0 z-30">
-          {@render switcherMenu('rail')}
-        </div>
-      {/if}
-    </div>
+    <ProfileSwitcher
+      variant="compact"
+      query={sidebarNavigation.data}
+      {surface}
+      {loading}
+      bind:switcherOpen
+      {onProfileStateChanged}
+    />
 
     <nav class="flex w-full flex-col items-center gap-1" aria-label="주요 메뉴">
       {#each navItems as item}
@@ -500,16 +241,16 @@
         ></div>
       </div>
 
-      {#if sidebarActiveProfile}
-        {@const activeProfile = sidebarActiveProfile}
-        <div class="absolute left-5 top-[54px] flex w-[280px] items-start justify-between">
-          <div
-            class="flex size-24 items-center justify-center rounded-full bg-[#111111] text-4xl font-bold text-white shadow-[1px_1px_2px_rgba(0,0,0,0.25)]"
-          >
-            {getProfileInitial(activeProfile.displayName, activeProfile.handle)}
-          </div>
-        </div>
+      <div class="absolute left-5 top-[54px] flex w-[280px] items-start justify-between">
+        <Avatar
+          class="size-24 text-4xl shadow-[1px_1px_2px_rgba(0,0,0,0.25)]"
+          initials={sidebarActiveProfile
+            ? getProfileInitial(sidebarActiveProfile.displayName, sidebarActiveProfile.handle)
+            : '?'}
+        />
+      </div>
 
+      {#if sidebarActiveProfile}
         <button
           class="absolute right-5 top-[134px] inline-flex h-8 items-center justify-center rounded-lg bg-[#fce79a] px-3 text-sm font-bold text-[#111111] opacity-60"
           type="button"
@@ -518,106 +259,41 @@
         >
           편집
         </button>
+      {/if}
 
-        <div class="absolute left-2.5 top-[140px] flex w-[300px] flex-col items-start px-2 py-2">
-          <button
-            class="flex h-[42px] max-w-full items-center gap-2 py-[5px] text-left"
-            type="button"
-            aria-label="프로필 목록"
-            aria-expanded={switcherOpen}
-            onclick={openProfileSwitcher}
-          >
-            <span class="truncate text-2xl font-bold leading-[32px] text-black/85"
-              >{activeProfile.displayName}</span
-            >
-            <svg
-              class="size-4 shrink-0 text-black/45"
-              viewBox="0 0 16 16"
-              fill="currentColor"
-              aria-hidden="true"
-            >
-              <path d="M4.2 6.2 8 10l3.8-3.8 1.1 1.1L8 12.2 3.1 7.3z" />
-            </svg>
-          </button>
+      <div class="absolute left-2.5 top-[140px] flex w-[300px] flex-col items-start px-2 py-2">
+        <ProfileSwitcher
+          variant="full"
+          query={sidebarNavigation.data}
+          {surface}
+          {loading}
+          bind:switcherOpen
+          {onProfileStateChanged}
+        />
+        {#if sidebarActiveProfile}
           <p class="max-w-full truncate text-sm leading-[19.6px] text-[#777777]">
-            @{activeProfile.handle}
+            @{sidebarActiveProfile.handle}
           </p>
           <div class="mt-2 flex items-center gap-3 text-sm leading-[22px] text-black">
             <span class="flex items-center gap-2 px-1"
-              ><span>{formatCount(activeProfile.followersCount ?? 0)}</span><span>팔로워</span
+              ><span>{formatCount(sidebarActiveProfile.followersCount ?? 0)}</span><span
+                >팔로워</span
               ></span
             >
             <span class="flex items-center gap-2 px-1"
-              ><span>{formatCount(activeProfile.followingCount ?? 0)}</span><span>팔로잉</span
+              ><span>{formatCount(sidebarActiveProfile.followingCount ?? 0)}</span><span
+                >팔로잉</span
               ></span
             >
           </div>
-        </div>
-      {:else if sidebarNavigation.data && sidebarProfiles.length > 0}
-        <div class="absolute left-5 top-[54px] flex w-[280px] items-start justify-between">
-          <div
-            class="flex size-24 items-center justify-center rounded-full bg-zinc-200 text-3xl font-bold text-zinc-500 shadow-[1px_1px_2px_rgba(0,0,0,0.25)]"
-          >
-            ?
-          </div>
-        </div>
-
-        <div class="absolute left-2.5 top-[140px] flex w-[300px] flex-col items-start px-2 py-2">
-          <button
-            class="flex h-[42px] max-w-full items-center gap-2 py-[5px] text-left"
-            type="button"
-            aria-label="프로필 목록"
-            aria-expanded={switcherOpen}
-            onclick={openProfileSwitcher}
-          >
-            <span class="truncate text-2xl font-bold leading-[32px] text-black/85">프로필 선택</span
-            >
-            <svg
-              class="size-4 shrink-0 text-black/45"
-              viewBox="0 0 16 16"
-              fill="currentColor"
-              aria-hidden="true"
-            >
-              <path d="M4.2 6.2 8 10l3.8-3.8 1.1 1.1L8 12.2 3.1 7.3z" />
-            </svg>
-          </button>
+        {:else if hasProfiles}
           <p class="text-sm leading-[19.6px] text-[#777777]">사용할 프로필을 선택해주세요.</p>
-        </div>
-      {:else}
-        <div
-          class="absolute left-5 top-[54px] flex size-24 items-center justify-center rounded-full bg-zinc-200 text-3xl font-bold text-zinc-500 shadow-[1px_1px_2px_rgba(0,0,0,0.25)]"
-        >
-          ?
-        </div>
-        <div class="absolute left-2.5 top-[140px] flex w-[300px] flex-col px-2 py-2">
-          <button
-            class="flex h-[42px] max-w-full items-center gap-2 py-[5px] text-left"
-            type="button"
-            aria-label="프로필 목록"
-            aria-expanded={switcherOpen}
-            onclick={openProfileSwitcher}
-          >
-            <span class="truncate text-2xl font-bold leading-[32px] text-black/85">프로필</span>
-            <svg
-              class="size-4 shrink-0 text-black/45"
-              viewBox="0 0 16 16"
-              fill="currentColor"
-              aria-hidden="true"
-            >
-              <path d="M4.2 6.2 8 10l3.8-3.8 1.1 1.1L8 12.2 3.1 7.3z" />
-            </svg>
-          </button>
+        {:else}
           <p class="text-sm leading-[19.6px] text-[#777777]">
             {loading ? '프로필을 불러오는 중입니다.' : '새 프로필을 만들어 시작하세요.'}
           </p>
-        </div>
-      {/if}
-
-      {#if switcherOpen}
-        <div class="absolute left-4 top-[210px] z-10">
-          {@render switcherMenu('full')}
-        </div>
-      {/if}
+        {/if}
+      </div>
     </section>
 
     <section
@@ -737,29 +413,13 @@
     </section>
   </div>
 
-  <section class="sr-only" aria-label="프로필 전환 상태">
-    {#if profileActionLoading && profileCreationOpen}
-      프로필 생성 중
-    {/if}
-
-    {#if profileActionLoading && !profileCreationOpen}
-      전환 중
-    {/if}
-
+  <section class="sr-only" aria-label="프로필 상태">
     {#if loading}
       프로필을 불러오는 중입니다.
     {:else if error}
       프로필을 불러오지 못했습니다.
-    {:else if sidebarProfiles.length === 0}
+    {:else if !hasProfiles}
       사용 가능한 프로필이 없습니다.
-    {/if}
-
-    {#if profileError}
-      {profileError}
-    {/if}
-
-    {#if profileCreationError}
-      {profileCreationError}
     {/if}
   </section>
 </aside>
